ABMC de Altas-Bajas de Personas

+++ Desktop Application/formPersonaAccidentes.vb
@@ -46,7 +46,7 @@
 #End Region
 
 #Region "Load and Set Data"
-    Friend Sub RefreshData(Optional ByVal PositionIDAccidente As Integer = 0, Optional ByVal RestoreCurrentPosition As Boolean = False)
+    Friend Sub RefreshData(Optional ByVal PositionIDAccidente As Byte = 0, Optional ByVal RestoreCurrentPosition As Boolean = False)
         Me.Cursor = Cursors.WaitCursor
 
         textboxApellidoNombre.Text = mPersonaActual.ApellidoNombre
